# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.productRating:
    properties:
      rating_product:
        description: Users     int `json:"rating_user"`
        type: integer
      rating_value:
        type: integer
    type: object
  handlers.productReview:
    properties:
      review:
        type: string
      review_product:
        type: integer
      review_user:
        type: integer
      time:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: ecom application
  title: swagger documentation for the application
  version: "1.0"
paths:
...
  /invoice/{ID}:
    get:
      consumes:
      - application/json
      description: Generate a payment invoice for a specific order and download it
        as a PDF
      parameters:
      - description: Order ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: PDF file generated and downloaded successfully
          schema:
            type: json
        "500":
          description: Failed to fetch invoice data or generate PDF file
          schema:
            type: json
      summary: Generate payment invoice for a specific order
      tags:
      - User-Invoice
  /products/rating:
    post:
      consumes:
      - application/json
      description: Store or update rating for a product based on user input
      parameters:
      - description: Rating data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.productRating'
      produces:
      - application/json
      responses:
        "200":
          description: Thanks for rating
          schema:
            type: json
        "400":
          description: Failed to bind data or failed to update rating
          schema:
            type: json
      summary: Store or update rating for a product
      tags:
      - User-Rating
  /products/review:
    post:
      consumes:
      - application/json
      description: Store or update review for a product based on user input
      parameters:
      - description: Review data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.productReview'
      produces:
      - application/json
      responses:
        "200":
          description: Thanks for your review
          schema:
            type: json
        "401":
          description: Failed to bind data or failed to update review
          schema:
            type: json
      summary: Store or update review for a product
      tags:
      - User-Review
...
  /user/profile/order:
    get:
      consumes:
      - application/json
      description: View orders placed by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Orders listed successfully
          schema:
            type: json
        "401":
          description: Failed to fetch order data
          schema:
            type: json
      summary: View orders placed by user
      tags:
      - User-OrderManagement
  /user/profile/order/{ID}:
    patch:
      consumes:
      - application/json
      description: cancel order placed by the authenticated user
      parameters:
      - description: Order ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order canceled successfully
          schema:
            type: json
        "401":
          description: Failed to fetch order details
          schema:
            type: json
      summary: Cancel order placed by user
      tags:
      - User-OrderManagement
  /user/profile/orderdetails/{ID}:
    get:
      consumes:
      - application/json
      description: View order details placed by the authenticated user
      parameters:
      - description: Order ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order details listed successfully
          schema:
            type: json
        "401":
          description: Failed to fetch order details
          schema:
            type: json
      summary: View order details placed by user
      tags:
      - User-OrderManagement
...
swagger: "2.0"